+++ traits/src/control/control.rs
@@ -273,7 +273,7 @@ pub trait Control {
     fn get_gpio_states(&self) -> GpioPinArr<GpioState>;
     fn get_gpio_readings(&self) -> GpioPinArr<Result<bool, GpioReadError>>;
     fn get_adc_states(&self) -> AdcPinArr<AdcState>;
-    fn get_adc_readings(&self) -> AdcPinArr<Result<u8, AdcReadError>>;
+    fn get_adc_readings(&self) -> AdcPinArr<Result<u16, AdcReadError>>;
     fn get_timer_modes(&self) -> TimerArr<TimerMode>;
     fn get_timer_states(&self) -> TimerArr<TimerState>;
     fn get_pwm_states(&self) -> PwmPinArr<PwmState>;
